v0.4.0-beta2: entropy scanning fixes
