+++ src/__init__.py
@@ -1,4 +1,4 @@
-from .compression import compress_directory, compress_directory_legacy, set_worker_cap
+from .compression_module import compress_directory, compress_directory_legacy, set_worker_cap
 from .stats import CompressionStats, print_compression_summary, Spinner, LegacyCompressionStats
 from .config import get_cpu_info
 from .timer import PerformanceMonitor, TimingStats
